--- src/main/java/com/example/newMock/Controller/KafkaConfigs.java
@@ -1,56 +1,0 @@
-package com.example.newMock.Controller;
-
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.core.*;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class KafkaConfigs {
-
-    @Value("${kafka.bootstrapAddress}")
-    private String bootstrapAddress;
-
-    @Configuration
-    public class KafkaProducerConfig {
-
-        @Bean
-        public ProducerFactory<String, String> producerFactory() {
-            Map<String, Object> configProps = new HashMap<>();
-            configProps.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapAddress);
-            configProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG,StringSerializer.class);
-            configProps.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-            return new DefaultKafkaProducerFactory<>(configProps);
-        }
-
-        @Bean
-        public KafkaTemplate<String, String> kafkaTemplate() {
-            return new KafkaTemplate<>(producerFactory());
-        }
-    }
-
-    @Configuration
-    @EnableKafka
-    public class KafkaConsumer {
-
-        @Value("${spring.kafka.bootstrap-servers}")
-        private String bootstrapServers;
-
-        @Bean
-        public ConsumerFactory<String, String> consumerFactory() {
-            Map<String, Object> props = new HashMap<>();
-            props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG,bootstrapServers);
-            props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG,StringDeserializer.class);
-            props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
-            return new DefaultKafkaConsumerFactory<>(props);
-        }
-    }
-
-}
